Move rspDisasm so it can be used with the spimdisasm installation
